pyproject.toml
[tool.poetry]
name = "demo-dspy"
version = "0.1.0"
description = "Simple DSPy tool selection demo with Ollama"
authors = ["Demo Author"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
pandas = "^2.2.0"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
python-multipart = "^0.0.9"
psutil = "^7.0.0"
mcp = "^1.13.1"
fastmcp = "^2.11.3"
dspy-ai = "^3.0.2"
rich = "^14.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"